+++ aks_cluster/provider.tf
@@ -3,7 +3,7 @@ terraform {
 }
 
 provider "azurerm" {
-  version = "~> 2.0.0"
+  version = "~> 2.7.0"
   features {}
 }